Add logger utility for tracing function execution
Instead of using bare fmt.Println (etc.) calls, it's more convenient to
use the `log` package from the standard library in a way that lets many
or all messages be turned on or off in one place. The idea is that you
import the package with the custom logger, get a custom logger object,
and then send Println to the logger object.

When the logger object is "activated" it will print to stdout, otherwise
it discards the message. This is still a fairly naive approach but at
least it saves you having to comment and uncomment a bunch of lines
sprinkled throughout a function, now you can just write and trace and
write and trace and deactivate the logger when you're finished.